Apply more suggestions
