+++ test/e2e/testdata/httproute-routingtype.yaml
@@ -4,12 +4,12 @@ metadata:
   name: custom-proxy-config
   namespace: gateway-conformance-infra
 spec:
-  routingType: service
+  routingType: Service
 ---
 apiVersion: gateway.networking.k8s.io/v1
 kind: HTTPRoute
 metadata:
-  name: httproute-to-backend-fqdn
+  name: httproute-to-service
   namespace: gateway-conformance-infra
 spec:
   parentRefs:
@@ -18,18 +18,7 @@ spec:
   - matches:
     - path:
         type: PathPrefix
-        value: /backend-fqdn
+        value: /to-cluster-ip
     backendRefs:
       - name: infra-backend-v1
         port: 8080
----
-apiVersion: gateway.envoyproxy.io/v1alpha1
-kind: Backend
-metadata:
-  name: backend-fqdn
-  namespace: gateway-conformance-infra
-spec:
-  endpoints:
-  - fqdn:
-      hostname: infra-backend-v1.gateway-conformance-infra.svc.cluster.local
-      port: 8080
